chore: Update docs with with the docs about the local executor
